apps/api/pyproject.toml
[project]
name = "chessminds-api"
version = "0.1.0"
description = "Chess of Minds — AWS Lambda handlers + shared engine specs"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-chess>=1.999,<2",
    "pydantic>=2.7,<3",
    "aws-lambda-powertools[tracer]>=3.0,<4",
    "boto3>=1.34,<2",
]

[dependency-groups]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "httpx>=0.27,<1",
    "moto[dynamodb]>=5,<6",
    "ruff>=0.6,<1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["engine", "handlers", "lib"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]
